refactor: fold emissions into mechanism — no separate namespace, parser, or file
Addresses K20shores' review of #281:
- Emissions types moved into mechanism_configuration::types (types.hpp)
- EmissionsConfig added as std::optional<types::EmissionsConfig> on Mechanism
- emissions: key parsed as an optional section by the main v1 Parse() call
- Standalone ParseEmissions() API and emissions::v1 namespace removed
- data_root and shell expansion dropped for JSON parity
- Examples updated: emissions.yaml gone; full_configuration.yaml gains an
  emissions: section
- Tests routed through Parse() instead of the old standalone parser
