+++ include/mechanism_configuration/mechanism.hpp
@@ -8,6 +8,7 @@
 
 #include <filesystem>
 #include <iostream>
+#include <optional>
 #include <string>
 #include <variant>
 
@@ -90,5 +91,7 @@ namespace mechanism_configuration
     Version version;
     /// @brief Relative tolerance for solver (optional, default: 1e-6)
     double relative_tolerance{ 1e-6 };
+    /// @brief Emissions configuration (optional)
+    std::optional<types::EmissionsConfig> emissions;
   };
 }  // namespace mechanism_configuration
