+++ src/CMakeLists.txt
@@ -43,7 +43,6 @@ target_include_directories(mechanism_configuration
 
 add_subdirectory(v0)
 add_subdirectory(v1)
-add_subdirectory(emissions)
 
 target_link_libraries(mechanism_configuration
   PUBLIC
