--- src/emissions/v1/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(mechanism_configuration
-  PRIVATE
-    ${CMAKE_CURRENT_SOURCE_DIR}/parser.cpp
-)
