+++ src/errors.cpp
@@ -34,7 +34,6 @@ namespace mechanism_configuration
       case ErrorCode::FileNotFound: return "FileNotFound";
       case ErrorCode::UnexpectedError: return "UnexpectedError";
       case ErrorCode::EmptyObject: return "EmptyObject";
-      case ErrorCode::NotAnEmissionsConfig: return "NotAnEmissionsConfig";
       case ErrorCode::DuplicateInventoryDetected: return "DuplicateInventoryDetected";
       case ErrorCode::DuplicateSpeciesMapDetected: return "DuplicateSpeciesMapDetected";
       case ErrorCode::DuplicateSourceDetected: return "DuplicateSourceDetected";
@@ -42,7 +41,6 @@ namespace mechanism_configuration
       case ErrorCode::SourceRequiresUnknownInventory: return "SourceRequiresUnknownInventory";
       case ErrorCode::SourceRequiresUnknownSpeciesMap: return "SourceRequiresUnknownSpeciesMap";
       case ErrorCode::SpeciesMapScalingExceedsOne: return "SpeciesMapScalingExceedsOne";
-      case ErrorCode::UndefinedEnvironmentVariable: return "UndefinedEnvironmentVariable";
       case ErrorCode::OnlineSourcesNotSupported: return "OnlineSourcesNotSupported";
       case ErrorCode::UnsupportedRegriddingType: return "UnsupportedRegriddingType";
       case ErrorCode::UnsupportedVerticalInjection: return "UnsupportedVerticalInjection";
