+++ test/unit/emissions/v1/emissions_unit_configs/duplicate_source.yaml
@@ -1,30 +1,32 @@
-kind: emissions
 version: 1.0.0
+species: []
+phases: []
+reactions: []
+emissions:
+  inventories:
+    cams bc:
+      directory: cams
+      file pattern: CAMS-GLOB-ANT_{YYYY}-{MM}.nc
+      convention: uptempo
 
-inventories:
-  cams bc:
-    directory: cams
-    file pattern: CAMS-GLOB-ANT_{YYYY}-{MM}.nc
-    convention: uptempo
+  species maps:
+    bc map:
+      mappings:
+        - inventory species: bc_anth_sum
+          mechanism species: BC
 
-species maps:
-  bc map:
-    mappings:
-      - inventory species: bc_anth_sum
-        mechanism species: BC
-
-sources:
-  - name: CAMS black carbon
-    mode: offline
-    type: anthropogenic
-    inventory: cams bc
-    species map: bc map
-    category: 0
-    hierarchy: 1
-  - name: CAMS black carbon
-    mode: offline
-    type: anthropogenic
-    inventory: cams bc
-    species map: bc map
-    category: 1
-    hierarchy: 1
+  sources:
+    - name: CAMS black carbon
+      mode: offline
+      type: anthropogenic
+      inventory: cams bc
+      species map: bc map
+      category: 0
+      hierarchy: 1
+    - name: CAMS black carbon
+      mode: offline
+      type: anthropogenic
+      inventory: cams bc
+      species map: bc map
+      category: 1
+      hierarchy: 1
