+++ test/unit/emissions/v1/emissions_unit_configs/scaling_exceeds_one.yaml
@@ -1,20 +1,22 @@
-kind: emissions
 version: 1.0.0
+species: []
+phases: []
+reactions: []
+emissions:
+  inventories:
+    cams bc:
+      directory: cams
+      file pattern: CAMS-GLOB-ANT_{YYYY}-{MM}.nc
+      convention: uptempo
 
-inventories:
-  cams bc:
-    directory: cams
-    file pattern: CAMS-GLOB-ANT_{YYYY}-{MM}.nc
-    convention: uptempo
+  species maps:
+    bad map:
+      mappings:
+        - inventory species: NOx
+          mechanism species: NO
+          scaling factor: 0.9
+        - inventory species: NOx
+          mechanism species: NO2
+          scaling factor: 0.5
 
-species maps:
-  bad map:
-    mappings:
-      - inventory species: NOx
-        mechanism species: NO
-        scaling factor: 0.9
-      - inventory species: NOx
-        mechanism species: NO2
-        scaling factor: 0.5
-
-sources: []
+  sources: []
