+++ test/unit/emissions/v1/emissions_unit_configs/unsupported_regridding.yaml
@@ -1,20 +1,22 @@
-kind: emissions
 version: 1.0.0
+species: []
+phases: []
+reactions: []
+emissions:
+  inventories:
+    cams bc:
+      directory: cams
+      file pattern: CAMS-GLOB-ANT_{YYYY}-{MM}.nc
+      convention: uptempo
 
-inventories:
-  cams bc:
-    directory: cams
-    file pattern: CAMS-GLOB-ANT_{YYYY}-{MM}.nc
-    convention: uptempo
+  species maps:
+    bc map:
+      mappings:
+        - inventory species: bc_anth_sum
+          mechanism species: BC
 
-species maps:
-  bc map:
-    mappings:
-      - inventory species: bc_anth_sum
-        mechanism species: BC
+  regridding:
+    type: scrip
+    weights file: /path/to/weights.nc
 
-regridding:
-  type: scrip
-  weights file: /path/to/weights.nc
-
-sources: []
+  sources: []
