+++ test/unit/emissions/v1/emissions_unit_configs/unsupported_vertical_injection.yaml
@@ -1,24 +1,26 @@
-kind: emissions
 version: 1.0.0
+species: []
+phases: []
+reactions: []
+emissions:
+  inventories:
+    cams bc:
+      directory: cams
+      file pattern: CAMS-GLOB-ANT_{YYYY}-{MM}.nc
+      convention: uptempo
 
-inventories:
-  cams bc:
-    directory: cams
-    file pattern: CAMS-GLOB-ANT_{YYYY}-{MM}.nc
-    convention: uptempo
+  species maps:
+    bc map:
+      mappings:
+        - inventory species: bc_anth_sum
+          mechanism species: BC
 
-species maps:
-  bc map:
-    mappings:
-      - inventory species: bc_anth_sum
-        mechanism species: BC
-
-sources:
-  - name: CAMS black carbon plume
-    mode: offline
-    type: fire
-    inventory: cams bc
-    species map: bc map
-    vertical injection: plume
-    category: 0
-    hierarchy: 1
+  sources:
+    - name: CAMS black carbon plume
+      mode: offline
+      type: fire
+      inventory: cams bc
+      species map: bc map
+      vertical injection: plume
+      category: 0
+      hierarchy: 1
