+++ test/unit/emissions/v1/test_parse_emissions.cpp
@@ -2,58 +2,53 @@
 //                         University of Illinois at Urbana-Champaign
 // SPDX-License-Identifier: Apache-2.0
 
-#include <mechanism_configuration/emissions/parse.hpp>
 #include <mechanism_configuration/errors.hpp>
+#include <mechanism_configuration/parse.hpp>
 
 #include <gtest/gtest.h>
 
 #include <filesystem>
 
 using namespace mechanism_configuration;
-namespace em = mechanism_configuration::emissions::v1;
 
-// Wrapper so string literals route to the file-path overload without ambiguity.
 static auto ParseFile(const char* p)
 {
-  return em::ParseEmissions(std::filesystem::path(p));
+  return Parse(std::filesystem::path(p));
 }
-using em::ParseEmissions;
 
 TEST(EmissionsV1Parser, ParsesValidConfig)
 {
   auto result = ParseFile("emissions_unit_configs/valid.yaml");
   ASSERT_TRUE(result) << "Parse failed: " << (result ? "" : result.error()[0].second);
 
-  const auto& config = *result;
-  EXPECT_EQ(config.name, "Valid test emissions config");
-  EXPECT_EQ(config.data_root, "/test/data");
-  EXPECT_EQ(config.version.major, 1u);
+  ASSERT_TRUE(result->emissions.has_value());
+  const auto& em = *result->emissions;
 
-  ASSERT_EQ(config.inventories.size(), 1u);
-  ASSERT_TRUE(config.inventories.count("cams bc"));
-  EXPECT_EQ(config.inventories.at("cams bc").directory, "cams");
-  EXPECT_EQ(config.inventories.at("cams bc").file_pattern, "CAMS-GLOB-ANT_{YYYY}-{MM}.nc");
-  EXPECT_EQ(config.inventories.at("cams bc").convention, "uptempo");
+  ASSERT_EQ(em.inventories.size(), 1u);
+  ASSERT_TRUE(em.inventories.count("cams bc"));
+  EXPECT_EQ(em.inventories.at("cams bc").directory, "cams");
+  EXPECT_EQ(em.inventories.at("cams bc").file_pattern, "CAMS-GLOB-ANT_{YYYY}-{MM}.nc");
+  EXPECT_EQ(em.inventories.at("cams bc").convention, "uptempo");
 
-  ASSERT_EQ(config.species_maps.size(), 1u);
-  ASSERT_TRUE(config.species_maps.count("bc map"));
-  const auto& smap = config.species_maps.at("bc map");
+  ASSERT_EQ(em.species_maps.size(), 1u);
+  ASSERT_TRUE(em.species_maps.count("bc map"));
+  const auto& smap = em.species_maps.at("bc map");
   ASSERT_EQ(smap.mappings.size(), 1u);
   EXPECT_EQ(smap.mappings[0].inventory_species, "bc_anth_sum");
   EXPECT_EQ(smap.mappings[0].mechanism_species, "BC");
   EXPECT_DOUBLE_EQ(smap.mappings[0].scaling_factor, 1.0);
 
-  EXPECT_EQ(config.regridding.type, em::types::RegriddingType::None);
+  EXPECT_EQ(em.regridding.type, types::RegriddingType::None);
 
-  ASSERT_EQ(config.sources.size(), 1u);
-  const auto& src = config.sources[0];
+  ASSERT_EQ(em.sources.size(), 1u);
+  const auto& src = em.sources[0];
   EXPECT_EQ(src.name, "CAMS black carbon");
-  EXPECT_EQ(src.mode, em::types::SourceMode::Offline);
-  EXPECT_EQ(src.type, em::types::SourceType::Anthropogenic);
+  EXPECT_EQ(src.mode, types::SourceMode::Offline);
+  EXPECT_EQ(src.type, types::SourceType::Anthropogenic);
   EXPECT_EQ(src.inventory, "cams bc");
   EXPECT_EQ(src.species_map, "bc map");
-  EXPECT_EQ(src.temporal_interpolation, em::types::TemporalInterpolation::Linear);
-  EXPECT_EQ(src.vertical_injection, em::types::VerticalInjection::Surface);
+  EXPECT_EQ(src.temporal_interpolation, types::TemporalInterpolation::Linear);
+  EXPECT_EQ(src.vertical_injection, types::VerticalInjection::Surface);
   EXPECT_EQ(src.category, 0);
   EXPECT_EQ(src.hierarchy, 1);
   EXPECT_DOUBLE_EQ(src.scaling_factor, 1.0);
@@ -65,58 +60,60 @@ TEST(EmissionsV1Parser, ParsesValidConfig)
 TEST(EmissionsV1Parser, ParsesValidConfigFromString)
 {
   const std::string content = R"(
-kind: emissions
 version: 1.0.0
-name: In-memory test
-
-inventories:
-  my inv:
-    directory: data
-    file pattern: file_{YYYY}.nc
-    convention: uptempo
-
-species maps:
-  my map:
-    mappings:
-      - inventory species: SO2
-        mechanism species: SO2
-
-sources:
-  - name: so2 source
-    mode: offline
-    type: anthropogenic
-    inventory: my inv
-    species map: my map
-    category: 0
-    hierarchy: 1
+species: []
+phases: []
+reactions: []
+emissions:
+  inventories:
+    my inv:
+      directory: data
+      file pattern: file_{YYYY}.nc
+      convention: uptempo
+  species maps:
+    my map:
+      mappings:
+        - inventory species: SO2
+          mechanism species: SO2
+  sources:
+    - name: so2 source
+      mode: offline
+      type: anthropogenic
+      inventory: my inv
+      species map: my map
+      category: 0
+      hierarchy: 1
 )";
 
-  auto result = ParseEmissions(content);
+  auto result = Parse(content);
   ASSERT_TRUE(result);
-  EXPECT_EQ(result->name, "In-memory test");
-  ASSERT_EQ(result->sources.size(), 1u);
-  EXPECT_EQ(result->sources[0].name, "so2 source");
+  ASSERT_TRUE(result->emissions.has_value());
+  ASSERT_EQ(result->emissions->sources.size(), 1u);
+  EXPECT_EQ(result->emissions->sources[0].name, "so2 source");
 }
 
-TEST(EmissionsV1Parser, RejectsWrongKind)
+TEST(EmissionsV1Parser, MechanismWithoutEmissionsHasNoEmissions)
 {
-  auto result = ParseFile("emissions_unit_configs/wrong_kind.yaml");
-  ASSERT_FALSE(result);
-  EXPECT_EQ(result.error()[0].first, ErrorCode::NotAnEmissionsConfig);
-}
-
-TEST(EmissionsV1Parser, RejectsMissingKind)
-{
-  auto result = ParseFile("emissions_unit_configs/missing_kind.yaml");
-  ASSERT_FALSE(result);
-  EXPECT_EQ(result.error()[0].first, ErrorCode::NotAnEmissionsConfig);
+  const std::string content = R"(
+version: 1.0.0
+species: []
+phases: []
+reactions: []
+)";
+  auto result = Parse(content);
+  ASSERT_TRUE(result);
+  EXPECT_FALSE(result->emissions.has_value());
 }
 
 TEST(EmissionsV1Parser, RejectsDuplicateSourceName)
 {
   auto result = ParseFile("emissions_unit_configs/duplicate_source.yaml");
   ASSERT_FALSE(result);
-  EXPECT_EQ(result.error()[0].first, ErrorCode::DuplicateSourceDetected);
+  bool found = false;
+  for (const auto& e : result.error())
+    if (e.first == ErrorCode::DuplicateSourceDetected)
+      found = true;
+  EXPECT_TRUE(found);
 }
 
 TEST(EmissionsV1Parser, RejectsDuplicateCategoryHierarchy)
@@ -134,65 +131,97 @@ TEST(EmissionsV1Parser, RejectsSourceWithUnknownInventory)
 {
   auto result = ParseFile("emissions_unit_configs/unknown_inventory.yaml");
   ASSERT_FALSE(result);
-  EXPECT_EQ(result.error()[0].first, ErrorCode::SourceRequiresUnknownInventory);
+  bool found = false;
+  for (const auto& e : result.error())
+    if (e.first == ErrorCode::SourceRequiresUnknownInventory)
+      found = true;
+  EXPECT_TRUE(found);
 }
 
 TEST(EmissionsV1Parser, RejectsSourceWithUnknownSpeciesMap)
 {
   auto result = ParseFile("emissions_unit_configs/unknown_species_map.yaml");
   ASSERT_FALSE(result);
-  EXPECT_EQ(result.error()[0].first, ErrorCode::SourceRequiresUnknownSpeciesMap);
+  bool found = false;
+  for (const auto& e : result.error())
+    if (e.first == ErrorCode::SourceRequiresUnknownSpeciesMap)
+      found = true;
+  EXPECT_TRUE(found);
 }
 
 TEST(EmissionsV1Parser, RejectsScalingFactorExceedingOne)
 {
   auto result = ParseFile("emissions_unit_configs/scaling_exceeds_one.yaml");
   ASSERT_FALSE(result);
-  EXPECT_EQ(result.error()[0].first, ErrorCode::SpeciesMapScalingExceedsOne);
+  bool found = false;
+  for (const auto& e : result.error())
+    if (e.first == ErrorCode::SpeciesMapScalingExceedsOne)
+      found = true;
+  EXPECT_TRUE(found);
 }
 
 TEST(EmissionsV1Parser, RejectsOnlineSource)
 {
   auto result = ParseFile("emissions_unit_configs/online_source.yaml");
   ASSERT_FALSE(result);
-  EXPECT_EQ(result.error()[0].first, ErrorCode::OnlineSourcesNotSupported);
+  bool found = false;
+  for (const auto& e : result.error())
+    if (e.first == ErrorCode::OnlineSourcesNotSupported)
+      found = true;
+  EXPECT_TRUE(found);
 }
 
 TEST(EmissionsV1Parser, RejectsUnsupportedRegriddingType)
 {
   auto result = ParseFile("emissions_unit_configs/unsupported_regridding.yaml");
   ASSERT_FALSE(result);
-  EXPECT_EQ(result.error()[0].first, ErrorCode::UnsupportedRegriddingType);
+  bool found = false;
+  for (const auto& e : result.error())
+    if (e.first == ErrorCode::UnsupportedRegriddingType)
+      found = true;
+  EXPECT_TRUE(found);
 }
 
 TEST(EmissionsV1Parser, RejectsUnsupportedVerticalInjection)
 {
   auto result = ParseFile("emissions_unit_configs/unsupported_vertical_injection.yaml");
   ASSERT_FALSE(result);
-  EXPECT_EQ(result.error()[0].first, ErrorCode::UnsupportedVerticalInjection);
+  bool found = false;
+  for (const auto& e : result.error())
+    if (e.first == ErrorCode::UnsupportedVerticalInjection)
+      found = true;
+  EXPECT_TRUE(found);
 }
 
 TEST(EmissionsV1Parser, AcceptsEmptySourcesList)
 {
   const std::string content = R"(
-kind: emissions
 version: 1.0.0
-sources: []
+species: []
+phases: []
+reactions: []
+emissions:
+  sources: []
 )";
-  auto result = ParseEmissions(content);
+  auto result = Parse(content);
   ASSERT_TRUE(result);
-  EXPECT_TRUE(result->sources.empty());
+  ASSERT_TRUE(result->emissions.has_value());
+  EXPECT_TRUE(result->emissions->sources.empty());
 }
 
 TEST(EmissionsV1Parser, AcceptsAbsentOptionalSections)
 {
   const std::string content = R"(
-kind: emissions
 version: 1.0.0
+species: []
+phases: []
+reactions: []
+emissions: {}
 )";
-  auto result = ParseEmissions(content);
+  auto result = Parse(content);
   ASSERT_TRUE(result);
-  EXPECT_TRUE(result->sources.empty());
-  EXPECT_TRUE(result->inventories.empty());
-  EXPECT_TRUE(result->species_maps.empty());
+  ASSERT_TRUE(result->emissions.has_value());
+  EXPECT_TRUE(result->emissions->sources.empty());
+  EXPECT_TRUE(result->emissions->inventories.empty());
+  EXPECT_TRUE(result->emissions->species_maps.empty());
 }
